Allow users to install Debian version of docker
An expert should know the different docker versions, 
and thus feel comfortable installing Debian's version instead.

Don't recommend this in the directions, but permit it to be done if they insist.  There can be reasons to prefer a distribution's package; such as the support guarantees or environments with restricted installs.